financial-ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.1
pydantic-settings==2.1.0
httpx==0.25.1
python-json-logger==2.0.7
google-generativeai==0.7.0
python-dotenv==1.0.0
